Move LeFigaro extractors to its own file
